初始化提交 登录 角色 权限 账号

+++ logic/RoleLogic.go
@@ -1,10 +1,10 @@
 package logic
 
 import (
-	"git.gumpcome.com/gumpoa/models"
-	"git.gumpcome.com/gumpoa/dao"
-	"git.gumpcome.com/gumpoa/constant"
-	"git.gumpcome.com/go_kit/dbkit"
+	"git.fenggese.com/go_kit/dbkit"
+	"git.fenggese.com/go_manage_system/constant"
+	"git.fenggese.com/go_manage_system/dao"
+	"git.fenggese.com/go_manage_system/models"
 )
 
 // @Title 添加角色
@@ -13,7 +13,7 @@ func AddRoleLogic(role *models.Role) (bool, int64, error) {
 	// 1. 检查参数
 	if role.Name == "" {
 		// 参数不合法
-		return false, 0, constant.RESP_CODE_PARAMS_ERROR
+		return false, 0, constant.RESP_CODE_PARAMS_VALUE_ERROR
 	}
 
 	// 2. 准备数据
@@ -31,12 +31,12 @@ func BindAccessLogic(roleAccess *models.RoleAccess) (bool, int64, error) {
 	// 1. 检查参数
 	if roleAccess.RoleID == 0 || roleAccess.AccessID == 0 {
 		// 参数不合法
-		return false, 0, constant.RESP_CODE_PARAMS_ERROR
+		return false, 0, constant.RESP_CODE_PARAMS_VALUE_ERROR
 	}
 
 	// 2. 准备数据
 	params := map[string]interface{}{
-		"role_id": roleAccess.RoleID,
+		"role_id":   roleAccess.RoleID,
 		"access_id": roleAccess.AccessID,
 	}
 
@@ -44,19 +44,18 @@ func BindAccessLogic(roleAccess *models.RoleAccess) (bool, int64, error) {
 	return dao.BindAccessDao(&params)
 }
 
-
 // @Title 解绑权限
 func UnbindAccessLogic(roleAccess *models.RoleAccess) (bool, error) {
 
 	// 1. 检查参数
 	if roleAccess.RoleID == 0 && roleAccess.AccessID == 0 {
 		// 参数不合法
-		return false, constant.RESP_CODE_PARAMS_ERROR
+		return false, constant.RESP_CODE_PARAMS_VALUE_ERROR
 	}
 
 	// 2. 准备数据
 	params := map[string]interface{}{
-		"role_id": roleAccess.RoleID,
+		"role_id":   roleAccess.RoleID,
 		"access_id": roleAccess.AccessID,
 	}
 
@@ -65,16 +64,37 @@ func UnbindAccessLogic(roleAccess *models.RoleAccess) (bool, error) {
 }
 
 // @Title 分页查找权限
-func PageRoleLogic(filter *models.PageRoleFilter) (dbkit.Page, error)  {
+func PageRoleLogic(filter *models.PageRoleFilter) (dbkit.Page, error) {
+	// 1. 检查参数
+	if filter.PageNumber < 1 || filter.PageSize < 0 {
+		// 参数不合法
+		return nil, constant.RESP_CODE_PARAMS_VALUE_ERROR
+	}
+
+	// 2. 分页查找
 	return dao.PageRoleDao(filter)
 }
 
 // @Title 查询角色的权限
 func AccessQueryByRoleLogic(role_id int64) ([]map[string]interface{}, error) {
+	// 1. 检查参数
+	if role_id <= 0 {
+		// 参数不合法
+		return nil, constant.RESP_CODE_PARAMS_VALUE_ERROR
+	}
+
+	// 2. 查询
 	return dao.AccessQueryByRoleDao(role_id)
 }
 
 // @Title 删除角色
 func DeleteRoleLogic(role_id int) (bool, error) {
+	// 1. 检查参数
+	if role_id <= 0 {
+		// 参数不合法
+		return nil, constant.RESP_CODE_PARAMS_VALUE_ERROR
+	}
+
+	// 2. 删除
 	return dao.DeleteRoleDao(role_id)
 }
